+++ Sources/EmojiKit/EmojiCategory+Persistency.swift
@@ -13,15 +13,70 @@ public extension EmojiCategory {
     /// A persisted list of emojis, that will be used by the
     /// ``EmojiCategory/favorites`` category.
     static var favoriteEmojis: [Emoji] {
-        get { getPersistedEmojis(for: "favorites") }
-        set { setPersistedEmojis(newValue, for: "favorites") }
+        get { getPersistedEmojis(for: .favorites) }
+        set { setPersistedEmojis(newValue, for: .favorites) }
     }
 
     /// A persisted list of emojis, that will be used by the
     /// ``EmojiCategory/frequent`` category.
     static var frequentEmojis: [Emoji] {
-        get { getPersistedEmojis(for: "frequent") }
-        set { setPersistedEmojis(newValue, for: "frequent") }
+        get { getPersistedEmojis(for: .frequent) }
+        set { setPersistedEmojis(newValue, for: .frequent) }
+    }
+}
+
+public extension EmojiCategory {
+
+    /// Add a certain emoji to a persisted category.
+    static func addEmoji(
+        _ emoji: Emoji,
+        to category: PersistedCategory,
+        maxCount: Int = 10_000
+    ) {
+        var emojis = getPersistedEmojis(for: category)
+            .filter { $0 != emoji }
+        emojis.insert(emoji, at: 0)
+        let result = Array(emojis.prefix(maxCount))
+        setPersistedEmojis(result, for: category)
+    }
+
+    /// Remove an emoji from a persisted category.
+    static func removeEmoji(
+        _ emoji: Emoji,
+        from category: PersistedCategory
+    ) {
+        let emojis = getPersistedEmojis(for: category)
+            .filter { $0 != emoji }
+        setPersistedEmojis(emojis, for: category)
+    }
+
+    /// Reset the emojis in a persisted category.
+    static func resetEmojis(
+        in category: PersistedCategory
+    ) {
+        setPersistedEmojis([], for: category)
+    }
+}
+
+public extension EmojiCategory {
+
+    /// This enum defines all standard, persisted categories.
+    enum PersistedCategory: Identifiable, Equatable {
+
+        case favorites, frequent, custom(name: String)
+    }
+}
+
+public extension EmojiCategory.PersistedCategory {
+
+    var id: String { name }
+
+    var name: String {
+        switch self {
+        case .favorites: "favorites"
+        case .frequent: "frequent"
+        case .custom(let name): name
+        }
     }
 }
 
@@ -29,7 +84,7 @@ extension EmojiCategory {
 
     /// Get a persisted list of emojis for a category.
     static func getPersistedEmojis(
-        for category: String
+        for category: PersistedCategory
     ) -> [Emoji] {
         let storage = persistedStorage
         let key = persistedStorageKey(for: category)
@@ -40,7 +95,7 @@ extension EmojiCategory {
     /// Get a persisted list of emojis for a category.
     static func setPersistedEmojis(
         _ emojis: [Emoji],
-        for category: String
+        for category: PersistedCategory
     ) {
         let storage = persistedStorage
         let key = persistedStorageKey(for: category)
@@ -59,8 +114,8 @@ extension EmojiCategory {
 
     /// Get the emojis storage key for custom category.
     static func persistedStorageKey(
-        for category: String
+        for category: PersistedCategory
     ) -> String {
-        "com.emojikit.category.\(category).emojis"
+        "com.emojikit.category.\(category.name).emojis"
     }
 }