+++ RELEASE_NOTES.md
@@ -15,6 +15,8 @@ This is the last minor version before the upcoming, first major version. The goa
 ### ✨ New Features
 
 * `EmojiCategory` now implements `CaseIterable`.
+* `EmojiCategory` has new functions to work with persisted categories.
+* `EmojiCategory.PersistedCategory` is a new enum to simplify working with persisted categories.
 * `EmojiCategory.favoriteEmojis` is a new value that can be used to populate the `.favorites` category.
 * `EmojiCategory.frequentEmojis` is a new value that can be used to populate the `.frequent` category.